Quiz Game
created a way includes the last time and selected option to patch the last condition to the db.json file due to continue to quiz from the last condition that user refresh the page or quit from the quiz. Fixed score calculation function. Made with of  progress bar is connected with current question. Going back on browser is linked to the categories page.