Quiz Game
Question types created and applied for options. For all types we have, created a way to get back the situation where the user stayed. Upgraded the score calculater for all types. Scoreboard updated for sorting the values alphabetically and for score point and also rank is created for the score values. Some of Result page components have finished for UI parts.